added @Valid anotation and Exceptions for invalid ID search

+++ Spring/taskSimpleBooks/library/src/main/java/lt/stud/techin/library/services/BookServiceImpl.java
@@ -1,6 +1,8 @@
 package lt.stud.techin.library.services;
 
+
 import lt.stud.techin.library.dataTransferObject.BookRequest;
+import lt.stud.techin.library.exceptions.BookByIdNotFoundException;
 import lt.stud.techin.library.model.Book;
 import lt.stud.techin.library.repository.BookRepository;
 import org.springframework.stereotype.Service;
@@ -11,6 +13,7 @@ import java.util.List;
 public class BookServiceImpl implements BookService{
 
     private final BookRepository bookRepository;
+    private Book book;
 
     public BookServiceImpl(BookRepository bookRepository) {
         this.bookRepository = bookRepository;
@@ -24,7 +27,7 @@ public class BookServiceImpl implements BookService{
     @Override
     public Book createBook(BookRequest newBook) {
 
-        Book book = new Book();
+        book = new Book();
 
         book.setAuthor(newBook.getAuthor());
         book.setTitle(newBook.getTitle());
@@ -32,7 +35,6 @@ public class BookServiceImpl implements BookService{
         book.setCover(newBook.getCover());
         book.setPrice(newBook.getPrice());
         book.setReserved(newBook.isReserved());
-
         bookRepository.save(book);
 
         return book;
@@ -40,8 +42,9 @@ public class BookServiceImpl implements BookService{
 
     @Override
     public Book updateBook(BookRequest newBook, Long id) {
-        return bookRepository.findById(id)
-                .map(book ->{
+        book = bookRepository.findById(id)
+                .orElseThrow(() -> new BookByIdNotFoundException("Book with ID: " + id + ", not found."));
+
                     book.setAuthor(newBook.getAuthor());
                     book.setTitle(newBook.getTitle());
                     book.setCategory(newBook.getCategory());
@@ -49,13 +52,18 @@ public class BookServiceImpl implements BookService{
                     book.setPrice(newBook.getPrice());
                     book.setReserved(newBook.isReserved());
 
-                    return bookRepository.save(book);
-                }).orElseThrow();
+                    bookRepository.save(book);
+
+                    return book;
     }
 
     @Override
     public void removeBook(Long id) {
+        if(bookRepository.existsById(id)) {
         bookRepository.deleteById(id);
+        } else {
+            throw new BookByIdNotFoundException("Book with ID: " + id + ", not found.");
+        }
     }
 
 }
